Cleaned up type/ldtk.odin
There was a lot of unneeded definitions in type/ldtk.odin, which weren't
used outside of the editor. Also a lot of things that are hard to
implement and are rare to be used. Removing them heavily boosts the
loading times. Stuff like BackgroundPosition or NineSlice should be
added back later